+++ packages/react-essentials-commands/src/_bin/middlewares/nextConfig.middleware.ts
@@ -1,8 +1,9 @@
 import { files, folders } from "#src/utils";
 
-import createMiddleware, { type Context } from "./createMiddleware";
+import createFileMiddleware from "./createFileMiddleware";
+import { type Context } from "./Middleware.types";
 
-const MIDDLEWARE = createMiddleware<string>({
+const MIDDLEWARE = createFileMiddleware<string>({
   path: "next.config.js",
   template: getTemplate,
   valid: ["app"],
@@ -11,16 +12,19 @@ const MIDDLEWARE = createMiddleware<string>({
 export default async function nextConfigMiddleware(
   context: Context,
 ): Promise<void> {
+  await Promise.all([MIDDLEWARE(context), deleteNextConfigFiles(context)]);
+  if (context.core === "app")
+    context.defer(() => folders.removeFolder(".next"));
+  else context.defer(() => files.removeFile("next-env.d.ts"));
+}
+
+async function deleteNextConfigFiles(context: Context): Promise<void> {
+  if (context.command !== "regenerate") return;
   await Promise.all([
-    MIDDLEWARE(context),
     files.removeFile("next.config.mjs"),
     files.removeFile("next.config.ts"),
     files.removeFile("out"),
   ]);
-
-  if (context.core !== "app")
-    context.defer(() => files.removeFile("next-env.d.ts"));
-  else context.defer(() => folders.removeFolder(".next"));
 }
 
 function getTemplate(context: Context): string {
