+++ packages/react-essentials-commands/src/_bin/middlewares/postCssConfig.middleware.ts
@@ -1,8 +1,9 @@
 import { files } from "#src/utils";
 
-import createMiddleware, { type Context } from "./createMiddleware";
+import createFileMiddleware from "./createFileMiddleware";
+import { type Context } from "./Middleware.types";
 
-const MIDDLEWARE = createMiddleware<string>({
+const MIDDLEWARE = createFileMiddleware<string>({
   path: "postcss.config.js",
   template: getTemplate,
   valid: ["app", "lib"],
@@ -11,8 +12,12 @@ const MIDDLEWARE = createMiddleware<string>({
 export default async function postCssConfigMiddleware(
   context: Context,
 ): Promise<void> {
+  await Promise.all([MIDDLEWARE(context), deletePostCSSConfigFiles(context)]);
+}
+
+async function deletePostCSSConfigFiles(context: Context): Promise<void> {
+  if (context.command !== "regenerate") return;
   await Promise.all([
-    MIDDLEWARE(context),
     files.removeFile("postcss.config.mjs"),
     files.removeFile("postcss.config.ts"),
   ]);
